SymfonyDZ3 v2

+++ api/src/Controller/OrdersController.php
@@ -14,6 +14,7 @@ use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
 use DateTimeImmutable;
+use Symfony\Component\Security\Core\Exception\AccessDeniedException;
 
 class OrdersController extends AbstractController
 {
@@ -41,7 +42,6 @@ class OrdersController extends AbstractController
     {
         $requestData = json_decode($request->getContent(), true);
         if (!isset(
-            $requestData['order_sum'],
             $requestData['car_ids'],
             $requestData['user_id']
         )) {
@@ -52,20 +52,28 @@ class OrdersController extends AbstractController
         if (!$user) {
             throw new Exception("User with id " . $requestData['user_id'] . " not found");
         }
+
+        $currentUser = $this->getUser();
+        if ($currentUser !== $user || !in_array(User::ROLE_USER, $currentUser->getRoles())) {
+            throw new AccessDeniedException('Unauthorized');
+        }
+
         $timezone = new DateTimeZone('Europe/Kiev');
         $orderDate = new DateTimeImmutable('now', $timezone);
         $order = new Orders();
         $order
             ->setOrderDate($orderDate)
-            ->setOrderSum($requestData['order_sum'])
             ->setUser($user);
+        $totalOrderSum = 0;
         foreach ($requestData['car_ids'] as $carId) {
             $car = $this->entityManager->getRepository(Car::class)->find($carId);
             if (!$car) {
                 throw new Exception("Car with id " . $carId . " not found");
             }
             $order->addCar($car);
+            $totalOrderSum += $car->getPrice();
         }
+        $order->setOrderSum($totalOrderSum);
         $this->entityManager->persist($order);
         $this->entityManager->flush();
 
@@ -78,6 +86,7 @@ class OrdersController extends AbstractController
     #[Route('order-all', name: 'order_all')]
     public function getAll(): JsonResponse
     {
+        $this->checkAdminAuthorization();
         $order = $this->entityManager->getRepository(Orders::class)->findAll();
 
         return new JsonResponse($order);
@@ -91,10 +100,15 @@ class OrdersController extends AbstractController
     #[Route('order/{id}', name: 'order_get_item')]
     public function getOrder(string $id): JsonResponse
     {
+        $currentUser = $this->getUser();
         /** @var Orders $order */
         $order = $this->entityManager->getRepository(Orders::class)->find($id);
         if (!$order) {
             throw new Exception("Order with id " . $id . " not found");
+        }
+
+        if ($currentUser !== $order->getUser()) {
+            throw new AccessDeniedException('Unauthorized');
         }
 
         return new JsonResponse($order);
@@ -108,10 +122,15 @@ class OrdersController extends AbstractController
     #[Route('order-update/{id}', name: 'order_update_item')]
     public function updateOrder(string $id): JsonResponse
     {
+        $currentUser = $this->getUser();
         /** @var Orders $order */
         $order = $this->entityManager->getRepository(Orders::class)->find($id);
         if (!$order) {
             throw new Exception("Order with id " . $id . " not found");
+        }
+
+        if ($currentUser !== $order->getUser()) {
+            throw new AccessDeniedException('Unauthorized');
         }
 
         $order->setOrderSum(21000);
@@ -128,10 +147,15 @@ class OrdersController extends AbstractController
     #[Route('order-delete/{id}', name: 'order_delete_item')]
     public function deleteOrder(string $id): JsonResponse
     {
+        $currentUser = $this->getUser();
         /** @var Orders $order */
         $order = $this->entityManager->getRepository(Orders::class)->find($id);
         if (!$order) {
             throw new Exception("Order with id " . $id . " not found");
+        }
+
+        if ($currentUser !== $order->getUser()) {
+            throw new AccessDeniedException('Unauthorized');
         }
 
         $this->entityManager->remove($order);
@@ -139,5 +163,17 @@ class OrdersController extends AbstractController
 
         return new JsonResponse();
     }
+
+    /**
+     * @return void
+     */
+    public function checkAdminAuthorization(): void
+    {
+        $user = $this->getUser();
+        if (!$user || !in_array(User::ROLE_ADMIN, $user->getRoles())) {
+            throw new AccessDeniedException('Unauthorized');
+        }
+    }
+
 
 }
